Simplifying ban list tabs
